pyproject.toml
[project]
name = "company-research"
version = "0.1.0"
description = "転職活動のための企業調査ダッシュボードCLI"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "typer>=0.12",
    "rich>=13",
    "jinja2>=3.1",
    "pydantic>=2.7",
    "pymupdf>=1.24",
    "pillow>=10",
    "pillow-heif>=0.18",
    "python-dotenv>=1.0",
    "anyio>=4.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.7",
]

[project.scripts]
company-research = "research.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/research"]
